Port the load(path); function from Fiji's Javascript interpreter
The 'load(path);' function can be used to store libraries of Javascript
code in a central location and reuse them from multiple Javascripts.

It was noted by Christophe Leterrier that this support was not yet
available in SciJava's Javascript support.

The original code was written by yours truly, hence it is easy to
relicense it to BSD as required for this project.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>