WIP
This branch will only be necessary if we decide to explicitly ship the
Rhino script engine version of JavaScript, as opposed to leaning on
Nashorn's compatibility functions in Java 8.

Time will tell whether the Nashorn backwards compatibility is good
enough. But in the meantime, let's leave this branch for a while.